get some handles by making a board model
