+++ app/templates/components/board.hbs
@@ -1,5 +1,5 @@
 <div class="board" style="width:{{this.width}}px;height:{{this.height}}px">
-  {{#each board as |row ri|}}
+  {{#each board.board as |row ri|}}
     {{#each row as |col ci|}}
       <div
         class="block {{if col "active"}}"
